+++ statham/dsl/elements/__init__.py
@@ -2,6 +2,7 @@ from statham.dsl.elements.array import Array
 from statham.dsl.elements.base import Element
 from statham.dsl.elements.boolean import Boolean
 from statham.dsl.elements.composition import AnyOf, CompositionElement, OneOf
+from statham.dsl.elements.meta import ObjectOptions
 from statham.dsl.elements.null import Null
 from statham.dsl.elements.numeric import Integer, Number
 from statham.dsl.elements.object import Object