+++ statham/dsl/elements/meta.py
@@ -1,10 +1,42 @@
-from typing import Any, Dict, Tuple, Type
+from typing import Any, cast, Dict, Optional, Tuple, Type, Union
 
 from statham.dsl import validators as val
 from statham.dsl.elements.base import Element
+from statham.dsl.helpers import custom_repr_args
 from statham.dsl.property import _Property
 from statham.dsl.constants import NotPassed
 from statham.dsl.exceptions import SchemaDefinitionError
+
+
+class ObjectOptions:
+
+    additionalProperties: Optional[Element]
+
+    def __init__(self, *, additionalProperties: Union[Element, bool] = True):
+        # Name used to match JSONSchema.
+        # pylint: disable=invalid-name
+        if additionalProperties is False:
+            self.additionalProperties = None
+        elif additionalProperties is True:
+            self.additionalProperties = Element()
+        else:
+            self.additionalProperties = cast(Element, additionalProperties)
+
+    def __eq__(self, other) -> bool:
+        if not isinstance(other, ObjectOptions):
+            return False
+        return self.additionalProperties == other.additionalProperties
+
+    def __repr__(self):
+        args = custom_repr_args(self)
+        if args.kwargs["additionalProperties"] == Element():
+            del args.kwargs["additionalProperties"]
+        if args.kwargs["additionalProperties"] is None:
+            args.kwargs["additionalProperties"] = False
+        return f"{type(self).__name__}{repr(args)}"
+
+
+RESERVED_PROPERTIES = dir(object) + ["default", "options", "properties"]
 
 
 class ObjectClassDict(dict):
@@ -20,15 +52,18 @@ class ObjectClassDict(dict):
         super().__init__(*args, **kwargs)
         self.properties = {}
         self.default = self.get("default", NotPassed())
+        self.options = self.get("options", ObjectOptions())
 
     def __setitem__(self, key, value):
+        if key in RESERVED_PROPERTIES and isinstance(value, _Property):
+            raise SchemaDefinitionError.reserved_attribute(key)
         if key == "default":
-            if isinstance(value, _Property):
-                raise SchemaDefinitionError.reserved_attribute("default")
             self.default = value
+        if key == "options":
+            self.options = value
         if isinstance(value, _Property):
             value.bind_name(key)
-            return self.properties.__setitem__(value.name or key, value)
+            return self.properties.__setitem__(key, value)
         return super().__setitem__(key, value)
 
 
@@ -41,6 +76,7 @@ class ObjectMeta(type, Element):
     """
 
     properties: Dict[str, _Property]
+    options: ObjectOptions
 
     @staticmethod
     def __subclasses__():
@@ -60,6 +96,7 @@ class ObjectMeta(type, Element):
         for property_ in cls.properties.values():
             property_.bind_class(cls)
         cls.default = classdict.default
+        cls.options = classdict.options
         return cls
 
     def __hash__(cls):
@@ -80,7 +117,11 @@ class ObjectMeta(type, Element):
         super_cls = next(iter(cls.mro()[1:]))
         class_def = f"""class {repr(cls)}({super_cls.__name__}):
 """
-        if not cls.properties and isinstance(cls.default, NotPassed):
+        if (
+            not cls.properties
+            and isinstance(cls.default, NotPassed)
+            and cls.options == ObjectOptions()
+        ):
             class_def = (
                 class_def
                 + """
@@ -92,6 +133,12 @@ class ObjectMeta(type, Element):
                 class_def
                 + f"""
     default = {repr(cls.default)}
+"""
+            )
+        if cls.options != ObjectOptions():
+            class_def = class_def + (
+                f"""
+    options = {repr(cls.options)}
 """
             )
         for property_ in cls.properties.values():