+++ tests/dsl/elements/test_object.py
@@ -1,8 +1,9 @@
-from typing import List, Union
+from typing import Any, ClassVar, List, Union
 import pytest
 
 from statham.dsl.constants import Maybe, NotPassed
-from statham.dsl.elements import Array, Object, OneOf, String
+from statham.dsl.elements import Array, Integer, Object, OneOf, String
+from statham.dsl.elements.meta import ObjectOptions
 from statham.dsl.property import Property
 from statham.dsl.exceptions import SchemaDefinitionError, ValidationError
 from tests.helpers import no_raise
@@ -25,8 +26,10 @@ class ObjectWrapper(Object):
     obj: StringWrapper = Property(StringWrapper, required=True)
 
 
-@pytest.mark.parametrize("param", [dict(value="foo")])
-def test_string_wrapper_vaccepts_valid_arguments(param):
+@pytest.mark.parametrize(
+    "param", [dict(value="foo"), dict(value="foo", other_value="anything")]
+)
+def test_string_wrapper_accepts_valid_arguments(param):
     with no_raise():
         _ = StringWrapper(param)
 
@@ -72,7 +75,6 @@ def test_list_wrapper_assigns_not_assed_correctly():
         dict(list_of_stuff=[{"value": 1}]),
         dict(list_of_stuff=["foo", {"value": 1}]),
         dict(list_of_stuff=[1, {"value": "foo"}]),
-        dict(not_a="property"),
     ],
 )
 def test_list_wrapper_fails_on_invalid_arguments(param):
@@ -128,47 +130,25 @@ def test_object_annotation():
 
 class TestRenamedProperties:
     class PropertyRename(Object):
-        default = {
-            "default": "string",
-            "self": "me",
-            "properties": "some properties",
-            "options": "some options",
-        }
 
-        _default = Property(String(), name="default")
-        _self = Property(String(), name="self")
-        _properties = Property(String(), name="properties")
-        _options = Property(String(), name="options")
+        default = {"default": "string"}
+        options = ObjectOptions(additionalProperties=False)
+
+        _default = Property(String(), source="default")
 
     def test_that_it_accepts_no_args(self):
         with no_raise():
             instance = self.PropertyRename()
-        assert instance.default == "string"
-        assert instance.self == "me"
-        assert instance.properties == "some properties"
-        assert instance.options == "some options"
+        assert instance._default == "string"
 
     def test_that_it_accepts_explicit_args(self):
         with no_raise():
-            instance = self.PropertyRename(
-                {
-                    "default": "another string",
-                    "self": "you",
-                    "properties": "other properties",
-                    "options": "other options",
-                }
-            )
-        assert instance.default == "another string"
-        assert instance.self == "you"
-        assert instance.properties == "other properties"
-        assert instance.options == "other options"
+            instance = self.PropertyRename({"default": "another string"})
+        assert instance._default == "another string"
 
-    @pytest.mark.parametrize(
-        "key", ["_default", "_properties", "_self", "_options"]
-    )
-    def test_that_it_fails_to_accept_outer_arg_names(self, key):
+    def test_that_it_fails_to_accept_outer_arg_name(self):
         with pytest.raises(ValidationError):
-            _ = self.PropertyRename({key: "string"})
+            _ = self.PropertyRename({"_default": "string"})
 
 
 class TestBadPropertyConflictErrors:
@@ -176,4 +156,56 @@ class TestBadPropertyConflictErrors:
         with pytest.raises(SchemaDefinitionError):
 
             class MyObject(Object):
-                default: str = Property(String())
+                default = Property(String())
+
+
+class TestAdditionalPropertiesAsElement:
+    class MyObject(Object):
+
+        options = ObjectOptions(additionalProperties=Integer())
+        value = Property(String())
+
+    def test_valid_instantiation(self):
+        with no_raise():
+            instance = self.MyObject({"value": "foo", "other_value": 3})
+        assert instance.value == "foo"
+        assert instance.additional_properties["other_value"] == 3
+
+    def test_additional_value_is_not_accepted_for_declared_value(self):
+        with pytest.raises(ValidationError):
+            _ = self.MyObject({"value": 1, "other_value": 3})
+
+    def test_bad_additional_value_is_not_accepted(self):
+        with pytest.raises(ValidationError):
+            _ = self.MyObject({"value": "foo", "other_value": "bad"})
+
+
+class TestAdditionalPropertiesAsTrue:
+    class MyObject(Object):
+
+        options = ObjectOptions(additionalProperties=True)
+        value = Property(String())
+
+    @pytest.mark.parametrize(
+        "additional_value", ["a string", 1, 3.4, None, True]
+    )
+    def test_values_are_accepted(self, additional_value: Any):
+        with no_raise():
+            instance = self.MyObject(
+                {"value": "foo", "other_value": additional_value}
+            )
+        assert instance.additional_properties["other_value"] == additional_value
+
+
+class TestAdditionalPropertiesAsFalse:
+    class MyObject(Object):
+
+        options = ObjectOptions(additionalProperties=False)
+        value = Property(String())
+
+    @pytest.mark.parametrize(
+        "additional_value", ["a string", 1, 3.4, None, True]
+    )
+    def test_no_extra_values_are_accepted(self, additional_value: Any):
+        with pytest.raises(ValidationError):
+            _ = self.MyObject({"value": "foo", "other_value": additional_value})