+++ tests/dsl/parser/test_parse_object.py
@@ -2,7 +2,8 @@ from typing import Any, Dict
 
 import pytest
 
-from statham.dsl.elements import Element, Object, String
+from statham.dsl.elements import Element, Object, ObjectOptions, String
+from statham.dsl.exceptions import SchemaParseError
 from statham.dsl.parser import parse_element
 from statham.dsl.property import Property
 
@@ -26,7 +27,19 @@ class ObjectWrapper(Object):
 class ObjectWithDefaultProp(Object):
     default = {"default": "a string"}
 
-    _default = Property(String(), name="default")
+    _default = Property(String(), source="default")
+
+
+class ObjectWithAdditionalPropElement(Object):
+    options = ObjectOptions(additionalProperties=String())
+
+
+class ObjectWithAdditionalPropTrue(Object):
+    options = ObjectOptions(additionalProperties=True)
+
+
+class ObjectWithAdditionalPropFalse(Object):
+    options = ObjectOptions(additionalProperties=False)
 
 
 @pytest.mark.parametrize(
@@ -83,9 +96,41 @@ class ObjectWithDefaultProp(Object):
             ObjectWithDefaultProp,
             id="with-property-named-default",
         ),
+        pytest.param(
+            {
+                "type": "object",
+                "title": "ObjectWithAdditionalPropElement",
+                "additionalProperties": {"type": "string"},
+            },
+            ObjectWithAdditionalPropElement,
+            id="with-additional-properties-schema",
+        ),
+        pytest.param(
+            {
+                "type": "object",
+                "title": "ObjectWithAdditionalPropTrue",
+                "additionalProperties": True,
+            },
+            ObjectWithAdditionalPropTrue,
+            id="with-additional-properties-true",
+        ),
+        pytest.param(
+            {
+                "type": "object",
+                "title": "ObjectWithAdditionalPropFalse",
+                "additionalProperties": False,
+            },
+            ObjectWithAdditionalPropFalse,
+            id="with-additional-properties-false",
+        ),
     ],
 )
 def test_parse_object_produces_expected_element(
     schema: Dict[str, Any], expected: Element
 ):
     assert parse_element(schema) == expected
+
+
+def test_parse_object_with_no_title_raises():
+    with pytest.raises(SchemaParseError):
+        parse_element({"type": "object"})