+++ tests/test_serializer.py
@@ -1,5 +1,7 @@
+import pytest
+
 from statham.dsl.constants import Maybe
-from statham.dsl.elements import Object, String
+from statham.dsl.elements import Object, ObjectOptions, String
 from statham.dsl.parser import parse
 from statham.dsl.property import Property
 from statham.serializer import _IMPORT_STATEMENTS, serialize_python
@@ -41,7 +43,7 @@ class Parent(Object):
 
     category: Category = Property(Category, required=True)
 
-    _default: Maybe[str] = Property(String(), name='default')
+    _default: Maybe[str] = Property(String(), source='default')
 
 
 class Other(Object):
@@ -71,6 +73,43 @@ def test_parse_and_serialize_schema_with_no_args():
         """class NoProps(Object):
 
     pass
+"""
+    )
+
+
+def test_parse_and_serialize_schema_with_additional_options_element():
+    schema = {
+        "type": "object",
+        "title": "StringContainer",
+        "additionalProperties": {"type": "string"},
+    }
+    assert serialize_python(*parse(schema)) == _IMPORT_STATEMENTS + (
+        """class StringContainer(Object):
+
+    options = ObjectOptions(additionalProperties=String())
+"""
+    )
+
+
+def test_parse_and_serialize_schema_with_option_dependencies():
+    schema = {
+        "type": "object",
+        "title": "StringWrapperContainer",
+        "additionalProperties": {
+            "type": "object",
+            "title": "StringWrapper",
+            "properties": {"value": {"type": "string"}},
+        },
+    }
+    assert serialize_python(*parse(schema)) == _IMPORT_STATEMENTS + (
+        """class StringWrapper(Object):
+
+    value: Maybe[str] = Property(String())
+
+
+class StringWrapperContainer(Object):
+
+    options = ObjectOptions(additionalProperties=StringWrapper)
 """
     )
 
@@ -106,5 +145,18 @@ def test_serialize_object_wrapper_object():
         == """class ObjectWrapper(Object):
 
     value: Maybe[StringWrapper] = Property(StringWrapper)
+"""
+    )
+
+
+@pytest.mark.parametrize("additional_properties", [String(), False])
+def test_serialize_object_with_additional_options(additional_properties):
+    class AdditionalPropObject(Object):
+        options = ObjectOptions(additionalProperties=additional_properties)
+
+    assert AdditionalPropObject.python() == (
+        f"""class AdditionalPropObject(Object):
+
+    options = ObjectOptions(additionalProperties={additional_properties})
 """
     )