+++ CHANGELOG.md
@@ -14,11 +14,14 @@ Types of changes are:
 
 ## [Unreleased]
 ### Added
+* Added `source` keyword argument to `Property`. This allows mapping of input
+  properties to differeing model attribute names.
+* Added support for `additionalProperties` keyword argument.
 * Added support for `default` as a property name in object schemas.
 * Added support for `self` as a property name in object schemas.
 
 ### Changed
-* Moved bulk to python serialization logic to DSL element methods.
+* Moved bulk of python serialization logic to DSL element methods.
 
 ## [0.2.0] - 2020-03-20
 ### Added